Make runbook (#99)
* pseudo code for test and beginning to write createIssue function

* create a runbook for runbook

* removing commented out tests and code for createIssue function...they are not ready yet

* removing commented out tests and code for createIssue function...they are not ready yet

* intended to make a runbook with no parse errors but failed so trying again

* pseudo code for test and beginning to write createIssue function

* create a runbook for runbook

* removing commented out tests and code for createIssue function...they are not ready yet

* removing commented out tests and code for createIssue function...they are not ready yet

* intended to make a runbook with no parse errors but failed so trying again

* create a runbook for runbook

removing commented out tests and code for createIssue function...they are not ready yet

removing commented out tests and code for createIssue function...they are not ready yet

* updating runbook again

* create a runbook for runbook

* create a runbook for runbook

* Correctly handle case where gitRefUrl from change API does not match github.com (#94)

This was previously throwing '<variable> is not a function or its return value is not iterable'
as String.prototype.match was returning null when there were no matches.

At least now we will get debug logs.

* Enable source-map-support on Lambdas (#93)

* Enable source-map-support babel plugin

See the serverless-webpack example repo: https://github.com/serverless-heaven/serverless-webpack/blob/21393845fb173a6f806b0bc2bee0be7daf0adc86/examples/babel-webpack-4/.babelrc.

* Package lambda functions individually with webpack

This is one of the main gains of using serverless-webpack as it leads to significantly reduced bundles, otherwise you're essentially just minifying all your functions
but still shipping all of them for each function.

* Move aws-sdk to devDependencies

This is automatically included in AWS Lambda environments so should not be shipped

* Remove individual packaging

Getting this error in logs:

        "TypeError: Cannot assign to read only property 'exports' of object '#<Object>'",
        "    at Module.config (/var/task/lambdas/ingester/src/webpack:/lambdas/ingester/src/lib/config.js:20:15)",
        "    at Module.key (/var/task/lambdas/ingester/src/webpack:/lambdas/ingester/src/lib/config.js:23:14)",
        "    at s (/var/task/lambdas/ingester/src/webpack:/webpack/bootstrap:9:1)",
        "    at Module.<anonymous> (/var/task/lambdas/ingester/src/webpack:/lambdas/ingester/src/templates/layout/layout.jsx:27:4)",
        "    at Module.<anonymous> (/var/task/lambdas/ingester/src/webpack:/lambdas/ingester/src/templates/layout/asset-loading.jsx:78:3)",
        "    at s (/var/task/lambdas/ingester/src/webpack:/webpack/bootstrap:9:1)",
        "    at Module.props (/var/task/lambdas/ingester/src/webpack:/lambdas/ingester/src/lib/render.jsx:7:18)",
        "    at Module.data-o-component (/var/task/lambdas/ingester/src/webpack:/lambdas/ingester/src/templates/layout/layout.jsx:21:5)",
        "    at s (/var/task/lambdas/ingester/src/webpack:/webpack/bootstrap:9:1)",
        "    at Module.<anonymous> (/var/task/lambdas/ingester/src/webpack:/external \"qs\":1:1)"

* Set babel-preset-env transform to commonjs explicitly

* Attempt to re-add individual packaging to serverless

This was not the root cause of previous errors, that was babel-preset-env.

* Use do-exclusively from rel-engage

* Log payload if no commit message was in the payload

* Correct the lint command in the README

* Use node10 on circle to match AWS Lambda

Lambda matches the latest node minor, so this keeps things more in sync

* Log change API Record in the event of the git commit info being missing

* Log github API call failures (#95)

* Log github API call failures

Currently Github API failures are logged as destructuring errors, as the response
is not in the expected array/object format.

This change propagates the child loggers to the github API caller, which logs an error and
throws on any non ok status code.

* Log parameters sent to GithubAPI on failure

* Redact github credentials from error logging

* Prevent logging of biz-ops API key

* Conditional api key deletion in logs

if headers does not exist this would currently throw

* Extract header API key name to variable

* Readme grammar improvements

* Log request method correctly

Everything is listed as a POST when it's actually a PATCH in some events

* Support valid github repository names (#97)

The current release log processing code doesn't support . or - characters in repository names. This results in incorrect parsing, e.g. runbook.md -> runbook.

* Log at start and end of release log ingestion (#96)

* Fix getting event IDs

* Use correct eventID field (#100)

Kinesis has eventID, not eventId. Nothing was being logged here.

* Add an integration test for one of the happy-paths for release-log (#98)

* Add an integration test for one of the happy-paths for release-log

There weren't any tests for this handler.

Added tests using nock to stub all API calls, and calling the handler with a mock kinesis record.

This is relatively brittle due to the nature of the calls, and also involves stubbing the schema using a 'private'ish API as the library doesn't expose the ability to use stub rawData - this should be a thing.

I've used the prometheus-pingdom-exporter runbook as a fixture for now
as it's relatively flushed out but this should probably be bespoke in
future.

Additionally moved the babel config to be shared between webpack
server/jest as jest blew up when getting near JSX.

* Rename runbook fixture so as to not be treated as a real runbook.md file

* RestoreAllMocks in other tests

* Do not enable source-map-support when running jest tests

Jest generates its own sourcemaps out-of-the-box. Enabling this plugin breaks
the line numbers of these sourcemaps.

* More explicit babel-preset-env node version

true is exactly equivalent to 'current' but the latter is much more obvious for the reader

* Fix jest.mock issue on node 12
https://github.com/facebook/jest/issues/8427

* Log eventIDs on completion

* Trace eventID through releaseLog flow

Trying to find missing events

* Log error on processing failure

* Add sourcemaps: both option to babel

This should hopefully reduce instances of 's is not a function' in the app logs

* Revert "Add sourcemaps: both option to babel"

This reverts commit 615b136913748368ff535c2624fda03f12fda9a1.
Babel loader already sets this option https://github.com/babel/babel-loader/blob/8d4353f3b8ba482212167231b9f8055fc9b64495/src/index.js

* Log awsRequestID on error

* Add start/stop debug logs at info level

* Log an explicit error when username is missing (#101)

There's an issue where if a users username in the FT github users repo doesn't
map case-sensitively to their github username then their FT username is not added to a change.

This is then not propagated to this service.

While we need to handle this, it's better to explicitly throw in this event for now rather than
failing when trying to split the string.

* removing parse error in runbook

* editing readme

* fixing merge conflicts in package lock